test-requirements.txt
pytest-it==0.1.4
pytest==7.2.2
sqlalchemy-utils==0.40.0
black==23.1.0
